Update Radio.pm with max Playlist lenght of 100
Since the Plugin stops adding tracks when a player is powered off and on again raising the Playlist Lenghth seems like a good idea. Restarting the Radio leads to many repeated tracks, so this is a workaround until somebody figues out a way to make the Plugin report played tracks and albums back to the server.